+++ docs/changelog.rst
@@ -7,6 +7,7 @@ Version 0.4.0 - Unreleased
 --------------------------
 
 * Enforce order on flat structure, close #6;
+* Added JSON structure, close #7;
 
 Version 0.3.0 - 2018/04/19
 --------------------------